Initial upload.
This is a test. There is some things that i will change later and think in better way to organize the documents. In later versions, on other repositories, i will create a "study room (tag, list, or whole repository) to upload the results of my studies and the method that i use to do it (through a lot of coments, that i love to write when studying - as you all see).